Actually make it an update to uvm-install2
