+++ install/uvm-install2/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "uvm-install2"
-version = "0.10.0"
+version = "0.11.0"
 authors = ["Manfred Endres <manfred.endres@tslarusso.de>"]
 description = "Install specified unity version."
 repository = "https://github.com/Larusso/unity-version-manager"
@@ -8,20 +8,25 @@ readme = "README.md"
 keywords = ["unity","version-manager"]
 categories = ["development-tools"]
 license = "Apache-2.0"
-edition = "2018"
+edition = "2021"
 
 [dependencies]
 console = "0.15.10"
+flexi_logger = "0.29.8"
 error-chain = "0.12.4"
 dirs-2 = "3.0.1"
-log = "0.4.26"
-structopt = "0.3.26"
-uvm_core = { path = "../../uvm_core", version = "0.13.3" }
-uvm_cli = { path = "../../uvm_cli", version = "0.4.0" }
-uvm_install_core = {version = "0.9.3" }
-uvm_install_graph = { version = "0.8.2" }
-
-[dev-dependencies]
-tempfile = "3.17.1"
-test-env-log = "0.2.8"
-env_logger = "0.11.6"
+indicatif = "0.17.11"
+log = "0.4.11"
+serde_json = "1.0.140"
+serde = "1.0"
+serde_derive = "1.0"
+thiserror = "2.0.11"
+clap = {  version = "4.5.29", features = ["derive", "string", "env", "cargo"] }
+unity-version = { path = "../../unity-version", version = "0.1.0"}
+uvm_live_platform = { path = "../../uvm_live_platform", version = "0.1.1"}
+uvm_paths = { path = "../../uvm_paths", version = "0.1.0"}
+uvm_install_graph = { path = "../../uvm_install_graph", version = "0.8.2"}
+unity-hub = {path = "../../unity-hub", version = "0.1.0" }
+uvm_install_core = {path = "../uvm_install_core" }
+anyhow = "1.0.75"
+lazy_static = "1.4.0"
