--- install/uvm-install3/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "uvm-install3"
-version = "0.1.0"
-authors = ["Manfred Endres <manfred.endres@tslarusso.de>"]
-description = "Install specified unity version."
-repository = "https://github.com/Larusso/unity-version-manager"
-readme = "README.md"
-keywords = ["unity","version-manager"]
-categories = ["development-tools"]
-license = "Apache-2.0"
-edition = "2021"
-
-[dependencies]
-console = "0.15.10"
-flexi_logger = "0.29.8"
-error-chain = "0.12.4"
-dirs-2 = "3.0.1"
-indicatif = "0.17.11"
-log = "0.4.11"
-serde_json = "1.0.140"
-serde = "1.0"
-serde_derive = "1.0"
-thiserror = "2.0.11"
-clap = {  version = "4.5.29", features = ["derive", "string", "env", "cargo"] }
-unity-version = { path = "../../unity-version", version = "0.1.0"}
-uvm_live_platform = { path = "../../uvm_live_platform", version = "0.1.1"}
-uvm_paths = { path = "../../uvm_paths", version = "0.1.0"}
-uvm_install_graph = { path = "../../uvm_install_graph", version = "0.8.2"}
-unity-hub = {path = "../../unity-hub", version = "0.1.0" }
-uvm_install_core = {path = "../uvm_install_core" }
-anyhow = "1.0.75"
-lazy_static = "1.4.0"
